Token count improvements; normalize \r chars; slightly better error status codes
The token counter now implements changes up to Pico-8 0.1.8, and may generally be more accurate as a result. I added a TODO for an unintentional edge case of "1..5" being one token, which Pico-8 is unlikely to ever fix (because it would increase counts on existing carts) so we'll need to support that at some point.

Some carts containing \r chars in source code were not parsing correctly because Pico-8 treats these as spaces while picotool was treating them as newlines. This only caused problems in line-sensitive cases such as short-if, and can only happen when code is edited by an external editor. The current fix is to replace "\r" with " " on loading, and this change will be saved to an updated/new cart. I'll wait until someone complains before doing something more thorough.

A few commands were not returning command line error status codes when encountering carts that didn't parse. Batch commands like "stats" do this when having partial success, which may not be the best choice. I've made a small change so that these commands report an error code on failure only when exactly one cart was given as input. This is not consistent with other commands that return an error code on partial success, so this deserves more thought later.

Some temp junk in this commit regarding .p8.png saving. Unfinished.