Compression routine, v1.
This compression routine is symmetrical with the decompression routine, with the exception of a newline that I add in the decomp routine that I forget why I'm doing that. (Added a TODO to find out why.) This currently does *not* reproduce Pico-8's compression result which (I think) is more efficient.

I also haven't checked yet whether the result of this comp is readable by Pico-8. One thing at a time. :)