Partial support for labels in .p8 files.
When loading a .p8 file with a __label__ section, the label is stored. When saving a cart with label data (loaded from a .p8 file or created fresh) as a .p8 file, the label is saved.

This does not yet convert labels between .p8 files and .p8.png files. Loading a .p8.png file will not populate the label data, and saving as a .p8.png will not use the label data (if any) in the label. (The original behavior is preserved: when saving as .p8.png, you can use another .p8.png file's label or use the default empty label.)