Support accidental if-do syntax.
Pico-8 accidentally supports this syntax:

```
if (cond) do
  ...
end
```

This is because it sees parentheses around the condition and the lack of "then" and assumes it's a short-if. During preprocessing, it replaces this with:

```
if (cond) then do end
  ...
  end
```

This accidentally does what the code author intended even though this is not valid Lua or intentional Pico-8 syntax.

A dozen or so carts have been published to the BBS that accidentally use this syntax. To parse these successfully, picotool pretends that if-do is a valid part of the language.