Cheap solution for luamin erroneously collapsing space between nonword char and a name.
The problem cases all appear to involve a name or keyword appearing after a closing bracket. It'd be more technically accurate to use the AST and make sure statements are preceded by a space. For the token-only minifier, this probably captures enough cases to fix the problem for carts where this matters.